Fix notification sent on proposal draft created, now sent on publish [backport 0.10]
