+++ decidim-proposals/app/commands/decidim/proposals/publish_proposal.rb
@@ -20,11 +20,40 @@ module Decidim
       #
       # Returns nothing.
       def call
-        return broadcast(:invalid) if @proposal.author != current_user
+        return broadcast(:invalid) if @proposal.author != @current_user
 
-        @proposal.update published_at: Time.current
+        transaction do
+          @proposal.update published_at: Time.current
+          send_notification
+          send_notification_to_participatory_space
+        end
 
         broadcast(:ok, @proposal)
+      end
+
+      private
+
+      def send_notification
+        return if @proposal.author.blank?
+
+        Decidim::EventsManager.publish(
+          event: "decidim.events.proposals.proposal_published",
+          event_class: Decidim::Proposals::PublishProposalEvent,
+          resource: @proposal,
+          recipient_ids: @proposal.author.followers.pluck(:id)
+        )
+      end
+
+      def send_notification_to_participatory_space
+        Decidim::EventsManager.publish(
+          event: "decidim.events.proposals.proposal_published",
+          event_class: Decidim::Proposals::PublishProposalEvent,
+          resource: @proposal,
+          recipient_ids: @proposal.participatory_space.followers.pluck(:id) - @proposal.author.followers.pluck(:id),
+          extra: {
+            participatory_space: true
+          }
+        )
       end
     end
   end
